Sidebar UI: Create WasmFileButton.svelte to handle .wasm/.wat file button logic
